Adjusted chan.queue_declare - named the queue from 'celery' to 'hypermap' - set passive to False
When adding it as a third party app to geonode I was getting the follwoing exception:

AMQPChannelException at /registry/celery_monitor/
(404, u"NOT_FOUND - no queue 'celery' in vhost '/'", (50, 10), 'Channel.queue_declare')

amqplib docs state that passive If set, the server will not create the queue. The client can use this to check whether a queue exists without modifying the server state.

Since the queue had not been created and was set to passive, the exception above was triggered.

Renamed the queue so it was not generalized as 'celery'.